# openapi.yaml
openapi: 3.0.3
info:
  title: Sunset Beach Resort & Spa — Admin/Booking API
  description: >
    Contract reverse-engineered from the current Next.js route handlers under
    `app/api/**/route.ts`. It documents exactly what each route accepts and
    returns today (including quirks such as Prisma `Decimal`/`Date` JSON
    serialization), so it can serve as a migration contract for a backend
    rewrite in another language. No behavior was changed to produce this spec.
  version: "1.0.0"
  license:
    name: Proprietary
  x-codegen-notes: >
    A property marked `x-sensitive: true` (guest email/phone, payment notes) is redacted as
    "[REDACTED]" in its generated model's toString() by the customized pojo.mustache template
    under openapi-templates/ - see scripts/generate-api.sh, which passes --template-dir. This is
    the one and only place PII redaction needs to be declared; it survives regeneration
    automatically. Do not hand-edit a generated model's toString() to redact a field instead -
    mark the field here and regenerate.
servers:
  - url: /api
    description: Relative to the deployed Next.js app origin.
...
security:
  - bearerAuth: []
paths:
...
  /printers:
    get:
      operationId: listPrinters
      tags: [Printers]
      summary: List all printers
      description: Requires role `MANAGER` or above.
      responses:
        "200":
          description: All printers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Printer"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenManagerPlus"
    post:
      operationId: createPrinter
      tags: [Printers]
      summary: Register a printer
      description: >
        Requires role `MANAGER` or above. Activating a printer (`isActive: true`, the default)
        for a `department` that already has an active printer is rejected with 409 - deactivate
        the old one first. Use `POST /printers/{id}/test` after creating one to confirm it's
        actually reachable before relying on it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrinterInput"
      responses:
        "201":
          description: Printer created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Printer"
        "400":
          description: Body failed validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenManagerPlus"
        "409":
          description: Another printer is already the active one for this `department`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: "Another printer is already active for department KITCHEN"
  /printers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updatePrinter
      tags: [Printers]
      summary: Update a printer
      description: >
        Requires role `MANAGER` or above. Full replacement of all `PrinterInput` fields - no
        partial update. Same 409-on-active-department-conflict rule as create.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrinterInput"
      responses:
        "200":
          description: Updated printer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Printer"
        "400":
          description: Body failed validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenManagerPlus"
        "404":
          description: Printer not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Printer not found
        "409":
          description: Another printer is already the active one for this `department`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: "Another printer is already active for department KITCHEN"
    delete:
      operationId: deletePrinter
      tags: [Printers]
      summary: Delete a printer
      description: >
        Requires role `MANAGER` or above. A printer that has ever printed anything (has
        `PrintJob` rows) can't be deleted - deactivate it (`isActive: false`) instead.
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkTrue"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenManagerPlus"
        "404":
          description: Printer not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Printer not found
        "409":
          description: This printer has existing print jobs and can't be deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: This printer has existing print jobs and can't be deleted.
  /printers/{id}/test:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: testPrinter
      tags: [Printers]
      summary: Print a test page
      description: >
        Requires role `MANAGER` or above. Attempts delivery immediately (bounded by a short
        socket timeout) and returns the resulting `PrintJob` either way - `status: SENT` means
        it printed, `status: PENDING`/`FAILED` means it didn't and `lastError` says why. This is
        the only way to actually confirm a printer's host/port/codepage are right; without it,
        setting one up is guesswork.
      responses:
        "201":
          description: Test print job created and attempted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintJob"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenManagerPlus"
        "404":
          description: Printer not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Printer not found
  /print-jobs:
    get:
      operationId: listPrintJobs
      tags: [Printers]
      summary: List print jobs
      description: >
        Requires role `WAITER` or above (i.e. any staff role) - but what comes back depends on
        who's asking. `ADMIN`/`MANAGER` see every document type. `CASHIER`/`WAITER` only see
        `KITCHEN_TICKET`/`BAR_TICKET`/`PREBILL` jobs - `Z_REPORT` (shift cash reconciliation) and
        `GUEST_RECEIPT` (booking/payment detail) are cashier/management information, filtered
        out rather than 403'd so line staff still get the one thing they actually need this
        endpoint for: seeing that their own kitchen/bar ticket or pre-bill didn't go through.
        Ordered by `createdAt` descending (most recent first). Filter by `status=FAILED` to see
        what needs manual attention - a silently lost kitchen ticket is a dish that never gets
        served - and/or by `documentType` to narrow to one kind of document (e.g.
        `documentType=Z_REPORT` for a manager auditing shift closes). Both filters apply
        server-side; an unrecognized value for either is a 400, not a silently-ignored no-op.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PrintJobStatus"
        - name: documentType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PrintDocumentType"
      responses:
        "200":
          description: Matching print jobs, filtered to the document types the caller's role may see.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrintJob"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /print-jobs/{id}/retry:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: retryPrintJob
      tags: [Printers]
      summary: Manually retry a print job
      description: >
        Requires role `WAITER` or above, subject to the same per-role document-type visibility
        as `GET /print-jobs`. Attempts delivery immediately regardless of the job's current
        `status` or `attempts` count (a human asking for a retry overrides the automatic-retry
        attempt ceiling). The original `payload` is replayed byte-for-byte - this re-sends
        exactly what was originally queued, it does not regenerate the document from current
        data. A job whose document type the caller's role can't see (e.g. a `WAITER` targeting a
        `Z_REPORT` job id) returns 404, not 403 - the id must not be usable to confirm the job
        exists, let alone force a reprint of it.
      responses:
        "200":
          description: 'The job after the retry attempt (`status: SENT` or `FAILED`).'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintJob"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Print job not found, or its document type isn't visible to the caller's role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Print job not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT returned by `POST /auth/login`, sent as `Authorization: Bearer
        <token>`. Contains `sub` (user id), `email`, `role`, standard `iat`/`exp`
        (default TTL 7 days, configurable via `JWT_TTL_DAYS`). Every route in
        this file except `POST /bookings` and the unauthenticated `/auth/*`
        endpoints (`login`, `logout`) requires a valid token, returning 401 if
        missing/invalid/expired, and 403 where the route additionally requires
        role `ADMIN`.
...
  schemas:
...
    Printer:
      type: object
      required: [id, name, department, host, port, codepage, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        department:
          $ref: "#/components/schemas/PrinterDepartment"
        host:
          type: string
          description: IP address or hostname on the hotel's local network.
        port:
          type: integer
          description: 'Almost always 9100 (raw ESC/POS-over-TCP) on real hardware.'
        codepage:
          $ref: "#/components/schemas/PrinterCodepage"
        isActive:
          type: boolean
          description: 'Whether this is currently the live printer for its `department` - see `Printer_one_active_per_department`.'
        createdAt:
          type: string
          format: date-time
...