# src/main/java/com/sunsetbeach/api/PrintersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sunsetbeach.api;

import com.sunsetbeach.model.ErrorMessage;
import com.sunsetbeach.model.OkTrue;
import com.sunsetbeach.model.PrintDocumentType;
import com.sunsetbeach.model.PrintJob;
import com.sunsetbeach.model.PrintJobStatus;
import com.sunsetbeach.model.Printer;
import com.sunsetbeach.model.PrinterInput;
import com.sunsetbeach.model.ValidationError;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.10.0")
@Validated
public interface PrintersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /printers : Register a printer
     * Requires role &#x60;MANAGER&#x60; or above. Activating a printer (&#x60;isActive: true&#x60;, the default) for a &#x60;department&#x60; that already has an active printer is rejected with 409 - deactivate the old one first. Use &#x60;POST /printers/{id}/test&#x60; after creating one to confirm it&#39;s actually reachable before relying on it. 
     *
     * @param printerInput  (required)
     * @return Printer created. (status code 201)
     *         or Body failed validation. (status code 400)
     *         or No valid JWT. (status code 401)
     *         or Token is valid but lacks the required role (&#x60;MANAGER&#x60; or above). (status code 403)
     *         or Another printer is already the active one for this &#x60;department&#x60;. (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/printers",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Printer> createPrinter(
         @Valid @RequestBody PrinterInput printerInput
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"port\" : 0, \"name\" : \"name\", \"host\" : \"host\", \"id\" : \"id\", \"department\" : \"KITCHEN\", \"codepage\" : \"PC437\", \"isActive\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"formErrors\" : [ ], \"fieldErrors\" : { \"guestEmail\" : [ \"Invalid email\" ] } } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /printers/{id} : Delete a printer
     * Requires role &#x60;MANAGER&#x60; or above. A printer that has ever printed anything (has &#x60;PrintJob&#x60; rows) can&#39;t be deleted - deactivate it (&#x60;isActive: false&#x60;) instead. 
     *
     * @param id  (required)
     * @return Deleted. (status code 200)
     *         or No valid JWT. (status code 401)
     *         or Token is valid but lacks the required role (&#x60;MANAGER&#x60; or above). (status code 403)
     *         or Printer not found. (status code 404)
     *         or This printer has existing print jobs and can&#39;t be deleted. (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/printers/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<OkTrue> deletePrinter(
         @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"ok\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /print-jobs : List print jobs
     * Requires role &#x60;WAITER&#x60; or above (i.e. any staff role) - but what comes back depends on who&#39;s asking. &#x60;ADMIN&#x60;/&#x60;MANAGER&#x60; see every document type. &#x60;CASHIER&#x60;/&#x60;WAITER&#x60; only see &#x60;KITCHEN_TICKET&#x60;/&#x60;BAR_TICKET&#x60;/&#x60;PREBILL&#x60; jobs - &#x60;Z_REPORT&#x60; (shift cash reconciliation) and &#x60;GUEST_RECEIPT&#x60; (booking/payment detail) are cashier/management information, filtered out rather than 403&#39;d so line staff still get the one thing they actually need this endpoint for: seeing that their own kitchen/bar ticket or pre-bill didn&#39;t go through. Ordered by &#x60;createdAt&#x60; descending (most recent first). Filter by &#x60;status&#x3D;FAILED&#x60; to see what needs manual attention - a silently lost kitchen ticket is a dish that never gets served - and/or by &#x60;documentType&#x60; to narrow to one kind of document (e.g. &#x60;documentType&#x3D;Z_REPORT&#x60; for a manager auditing shift closes). Both filters apply server-side; an unrecognized value for either is a 400, not a silently-ignored no-op. 
     *
     * @param status  (optional)
     * @param documentType  (optional)
     * @return Matching print jobs, filtered to the document types the caller&#39;s role may see. (status code 200)
     *         or No valid JWT. (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/print-jobs",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<PrintJob>> listPrintJobs(
         @Valid @RequestParam(value = "status", required = false) PrintJobStatus status,
         @Valid @RequestParam(value = "documentType", required = false) PrintDocumentType documentType
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"summary\" : \"summary\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"lastError\" : \"lastError\", \"documentType\" : \"KITCHEN_TICKET\", \"printerId\" : \"printerId\", \"id\" : \"id\", \"status\" : \"PENDING\", \"attempts\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }, { \"summary\" : \"summary\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"lastError\" : \"lastError\", \"documentType\" : \"KITCHEN_TICKET\", \"printerId\" : \"printerId\", \"id\" : \"id\", \"status\" : \"PENDING\", \"attempts\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /printers : List all printers
     * Requires role &#x60;MANAGER&#x60; or above.
     *
     * @return All printers. (status code 200)
     *         or No valid JWT. (status code 401)
     *         or Token is valid but lacks the required role (&#x60;MANAGER&#x60; or above). (status code 403)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/printers",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Printer>> listPrinters(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"port\" : 0, \"name\" : \"name\", \"host\" : \"host\", \"id\" : \"id\", \"department\" : \"KITCHEN\", \"codepage\" : \"PC437\", \"isActive\" : true }, { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"port\" : 0, \"name\" : \"name\", \"host\" : \"host\", \"id\" : \"id\", \"department\" : \"KITCHEN\", \"codepage\" : \"PC437\", \"isActive\" : true } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /print-jobs/{id}/preview : Preview a print job&#39;s rendered content
     * Requires role &#x60;WAITER&#x60; or above, subject to the same per-role document-type visibility as &#x60;GET /print-jobs&#x60; (404, not 403, for a job outside the caller&#39;s visible set - see that operation&#39;s description). Returns the same text that would come out of the printer, with the ESC/POS control sequences (init, code page selection, bold, centering, paper cut) stripped out - not a hex dump of &#x60;payload&#x60;. Useful both to check a document before it prints and to see what a &#x60;FAILED&#x60; job actually contained without walking over to a dead printer. 
     *
     * @param id  (required)
     * @return Plain-text rendering of the job&#39;s payload. (status code 200)
     *         or No valid JWT. (status code 401)
     *         or Print job not found, or its document type isn&#39;t visible to the caller&#39;s role. (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/print-jobs/{id}/preview",
        produces = { "text/plain", "application/json" }
    )
    
    default ResponseEntity<String> previewPrintJob(
         @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf(""))) {
                    String exampleString = "";
                    ApiUtil.setExampleResponse(request, "", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /print-jobs/{id}/retry : Manually retry a print job
     * Requires role &#x60;WAITER&#x60; or above, subject to the same per-role document-type visibility as &#x60;GET /print-jobs&#x60;. Attempts delivery immediately regardless of the job&#39;s current &#x60;status&#x60; or &#x60;attempts&#x60; count (a human asking for a retry overrides the automatic-retry attempt ceiling). The original &#x60;payload&#x60; is replayed byte-for-byte - this re-sends exactly what was originally queued, it does not regenerate the document from current data. A job whose document type the caller&#39;s role can&#39;t see (e.g. a &#x60;WAITER&#x60; targeting a &#x60;Z_REPORT&#x60; job id) returns 404, not 403 - the id must not be usable to confirm the job exists, let alone force a reprint of it. 
     *
     * @param id  (required)
     * @return The job after the retry attempt (&#x60;status: SENT&#x60; or &#x60;FAILED&#x60;). (status code 200)
     *         or No valid JWT. (status code 401)
     *         or Print job not found, or its document type isn&#39;t visible to the caller&#39;s role. (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/print-jobs/{id}/retry",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PrintJob> retryPrintJob(
         @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"summary\" : \"summary\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"lastError\" : \"lastError\", \"documentType\" : \"KITCHEN_TICKET\", \"printerId\" : \"printerId\", \"id\" : \"id\", \"status\" : \"PENDING\", \"attempts\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /printers/{id}/test : Print a test page
     * Requires role &#x60;MANAGER&#x60; or above. Attempts delivery immediately (bounded by a short socket timeout) and returns the resulting &#x60;PrintJob&#x60; either way - &#x60;status: SENT&#x60; means it printed, &#x60;status: PENDING&#x60;/&#x60;FAILED&#x60; means it didn&#39;t and &#x60;lastError&#x60; says why. This is the only way to actually confirm a printer&#39;s host/port/codepage are right; without it, setting one up is guesswork. 
     *
     * @param id  (required)
     * @return Test print job created and attempted. (status code 201)
     *         or No valid JWT. (status code 401)
     *         or Token is valid but lacks the required role (&#x60;MANAGER&#x60; or above). (status code 403)
     *         or Printer not found. (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/printers/{id}/test",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PrintJob> testPrinter(
         @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"summary\" : \"summary\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"lastError\" : \"lastError\", \"documentType\" : \"KITCHEN_TICKET\", \"printerId\" : \"printerId\", \"id\" : \"id\", \"status\" : \"PENDING\", \"attempts\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PATCH /printers/{id} : Update a printer
     * Requires role &#x60;MANAGER&#x60; or above. Full replacement of all &#x60;PrinterInput&#x60; fields - no partial update. Same 409-on-active-department-conflict rule as create. 
     *
     * @param id  (required)
     * @param printerInput  (required)
     * @return Updated printer. (status code 200)
     *         or Body failed validation. (status code 400)
     *         or No valid JWT. (status code 401)
     *         or Token is valid but lacks the required role (&#x60;MANAGER&#x60; or above). (status code 403)
     *         or Printer not found. (status code 404)
     *         or Another printer is already the active one for this &#x60;department&#x60;. (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/printers/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Printer> updatePrinter(
         @PathVariable("id") String id,
         @Valid @RequestBody PrinterInput printerInput
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"port\" : 0, \"name\" : \"name\", \"host\" : \"host\", \"id\" : \"id\", \"department\" : \"KITCHEN\", \"codepage\" : \"PC437\", \"isActive\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"formErrors\" : [ ], \"fieldErrors\" : { \"guestEmail\" : [ \"Invalid email\" ] } } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}